Rename main module #1
