+++ package.json
@@ -1,7 +1,8 @@
 {
   "scripts": {
     "lint": "flow",
-    "test": "jest --config jest.config.js --watch"
+    "test": "jest --config jest.config.js --watch",
+    "test:watch": "jest --config jest.config.js"
   },
   "devDependencies": {
     "babel-core": "^6.26.0",
