+++ src/utils/inferComponentName.js
